Add the header parse to xdp_lb as well
